Extract MyApp to own widget.
