+++ lib/main.dart
@@ -1,19 +1,4 @@
 import 'package:flutter/material.dart';
+import 'package:konjugator/widgets/my_app.dart';
 
 void main() => runApp(MyApp());
-
-class MyApp extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      home: Scaffold(
-        appBar: AppBar(
-          title: Text('Konjugator'),
-        ),
-        body: Center(
-          child: Text('Hello World'),
-        ),
-      ),
-    );
-  }
-}
